+++ openstack/blockstorage/v2/snapshots/util.go
@@ -8,7 +8,7 @@ import (
 // status. It will do this for the amount of seconds defined.
 func WaitForStatus(c *golangsdk.ServiceClient, id, status string, secs int) error {
 	return golangsdk.WaitFor(secs, func() (bool, error) {
-		current, err := Get(c, id).Extract()
+		current, err := Get(c, id)
 		if err != nil {
 			return false, err
 		}
@@ -19,4 +19,40 @@ func WaitForStatus(c *golangsdk.ServiceClient, id, status string, secs int) erro
 
 		return false, nil
 	})
+}
+
+// IDFromName is a convienience function that returns a snapshot's ID given its name.
+func IDFromName(client *golangsdk.ServiceClient, name string) (string, error) {
+	count := 0
+	id := ""
+
+	listOpts := ListOpts{
+		Name: name,
+	}
+
+	pages, err := List(client, listOpts).AllPages()
+	if err != nil {
+		return "", err
+	}
+
+	all, err := ExtractSnapshots(pages)
+	if err != nil {
+		return "", err
+	}
+
+	for _, s := range all {
+		if s.Name == name {
+			count++
+			id = s.ID
+		}
+	}
+
+	switch count {
+	case 0:
+		return "", golangsdk.ErrResourceNotFound{Name: name, ResourceType: "snapshot"}
+	case 1:
+		return id, nil
+	default:
+		return "", golangsdk.ErrMultipleResourcesFound{Name: name, Count: count, ResourceType: "snapshot"}
+	}
 }
